Moved SyncRecommendations to an Artisan command which is also run as scheduled task
