+++ ProcessMaker/Console/Commands/SyncRecommendationsCommand.php
@@ -2,8 +2,10 @@
 
 namespace ProcessMaker\Console\Commands;
 
+use Exception;
+use RuntimeException;
 use Illuminate\Console\Command;
-use ProcessMaker\SyncRecommendations;
+use Illuminate\Support\Facades\Http;
 
 class SyncRecommendationsCommand extends Command
 {
@@ -24,8 +26,67 @@ class SyncRecommendationsCommand extends Command
     /**
      * Execute the console command.
      */
-    public function handle()
+    public function handle(): void
     {
-        SyncRecommendations::sync();
+        $this->line('Syncing recommendations from GitHub...');
+
+        try {
+            /**
+             * WIP Need to continue business logic here to fetch the
+             * contents of each URL as JSON, parse it, check if we
+             * have existing recommendations (matching UUID) and
+             * if so, update them, or create new ones
+             */
+            $urls = static::remoteRecommendationsList();
+        } catch (Exception $e) {
+            $this->error($e->getMessage());
+        }
+    }
+
+    /**
+     * Returns an array of strings, each URL links to a JSON recommendation file.
+     *
+     * @return array
+     */
+    protected static function remoteRecommendationsList(): array
+    {
+        $response = Http::get(self::url('index.json'));
+
+        if ($response->failed()) {
+            throw new RuntimeException('Failed to retrieve recommendations from GitHub');
+        }
+
+        // JSON response is decoded into an array, where each key
+        // represents a top-level directory containing JSON files,
+        // each representing a recommendation
+        $directories = json_decode(file_get_contents($response->json()), true);
+
+        $urls = [];
+
+        foreach ($directories as $dir => $files) {
+            foreach ($files as $filename) {
+                $urls[] = self::url($dir.'/'.$filename);
+            }
+        }
+
+        return $urls;
+    }
+
+    /**
+     * Build the direct URL to the file/directory in the repo
+     *
+     * @param  string  $filename
+     *
+     * @return string
+     */
+    protected static function url(string $filename): string
+    {
+        $base = config('services.recommendations_github.base_url');
+
+        $repo = config('services.recommendations_github.repo');
+
+        $branch = config('services.recommendations_github.branch');
+
+        return "$base/$repo/$branch/$filename";
     }
 }
